+++ xdsl/dialects/arm/__init__.py
@@ -8,7 +8,7 @@ from typing import IO
 from xdsl.dialects.builtin import ModuleOp
 from xdsl.ir import Dialect
 
-from .ops import ARMOperation, DSMovOp, DSSMulOp, GetRegisterOp, RetOp
+from .ops import ARMOperation, DSMovOp, DSSMulOp, GetRegisterOp
 from .register import IntRegisterType
 
 
@@ -26,7 +26,6 @@ ARM = Dialect(
         GetRegisterOp,
         DSMovOp,
         DSSMulOp,
-        RetOp,
     ],
     [
         IntRegisterType,
